refactor : userRoutine과 routine 분리
