+++ src/main/java/com/focusup/domain/routine/dto/RoutineRequestDTO.java
@@ -11,23 +11,6 @@ import java.util.List;
 
 public class RoutineRequestDTO {
 
-    // 루틴 설정 request용 DTO
-    @Builder
-    @Getter
-    public static class CreateRoutine {
-
-        private String routineName;
-        private List<DayOfWeek> repeatCycleDay;
-
-        private LocalDate startDate;
-
-        @JsonFormat(shape = JsonFormat.Shape.STRING, pattern = "HH:mm", timezone = "Asia/Seoul")
-        private LocalTime startTime;
-
-        @JsonFormat(shape = JsonFormat.Shape.STRING, pattern = "HH:mm", timezone = "Asia/Seoul")
-        private LocalTime endTime;
-    }
-
     // 루틴 완료 request용 DTO
     @Getter
     public static class FinishRoutine {
