+++ src/main/java/com/focusup/domain/routine/dto/UserRoutineResponseDTO.java
@@ -1,0 +1,4 @@
+package com.focusup.domain.routine.dto;
+
+public class UserRoutineResponseDTO {
+}
