+++ src/main/java/com/focusup/domain/routine/repository/RoutineRepository.java
@@ -1,7 +1,6 @@
 package com.focusup.domain.routine.repository;
 
 import com.focusup.entity.Routine;
-import com.focusup.entity.User;
 import io.micrometer.common.lang.NonNullApi;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
