+++ src/main/java/com/focusup/domain/routine/service/RoutineService.java
@@ -2,18 +2,9 @@ package com.focusup.domain.routine.service;
 
 import com.focusup.domain.routine.dto.RoutineRequestDTO;
 import com.focusup.domain.routine.dto.RoutineResponseDTO;
-import com.focusup.entity.Routine;
-import com.focusup.entity.UserRoutine;
-
 import java.time.LocalDate;
-import java.time.LocalDateTime;
-import java.time.LocalTime;
 
 public interface RoutineService {
-    public Long createUserRoutine(RoutineRequestDTO.CreateRoutine request, Long userId);
-
-    public Routine createRoutineInfo(LocalDateTime startTime, LocalDateTime endTime, UserRoutine userRoutine);
-
     // 모든 루틴 리스트 조회 service
     public RoutineResponseDTO.GetAllRoutineList getAllRoutineList();
 
@@ -23,6 +14,4 @@ public interface RoutineService {
     // 루틴 완료 service
     public Long finishRoutine(RoutineRequestDTO.FinishRoutine request, Long routineId);
 
-    // 루틴 삭제 service
-    public void deleteRoutine(Long userRoutineId);
 }
